zero-shot-object-detector/src/requirements.txt
groundingdino-py==0.4.0
uvicorn==0.19.0
fastapi==0.85.0
python-multipart==0.0.5
numpy==1.24.2
prometheus-fastapi-instrumentator==6.0.0
Pillow==9.5.0
opencv-python-headless==4.7.0.72